IntではなくLongを受けるべき

+++ webapp/scala/app/controllers/Ticket.scala
@@ -3,6 +3,6 @@ import play.api._
 import play.api.mvc._
 
 object Ticket extends Controller {
-  def show(ticket_id: Int) = TODO
+  def show(ticket_id: Long) = TODO
   def buy() = TODO
 }
